requirements.txt
h5py==2.10.0
pandas==0.25.3
numpy==1.18.0
Pillow==6.2.1
torch==1.3.1
torchvision==0.4.2
tornado==5.1.1
